[storage] move FSStoragePlugin::getObjectFormatByExtension() to StorageItem
Some methods that are now in FSStoragePlugin logically rather belong to
StorageItem class, which was so far treated just as a simple data
container. Moving them will help to make the code better arranged and
simplify the convoluted storage plugin class.

This is the first in likely larger set of commits that split
responsibilities between FSStoragePlugin and StorageItem.